_alternate_id should return a server-side name
Resource._alternate_id currently returns the attribute on the Resource
class that acts as an alternate id when it actually needs to return the
name that's associated with that attribute internally, i.e., the
Body attribute that it's connected to.

Glance is documented incorrectly and works inconsistently between using
"member_id" and "member" fields for the Member resource, and
_alternate_id returning "member_id" breaks Member.list. It tries to
search for "member_id" in the JSON when it's told to look for that
alternate, but we should really be telling it not that Member.member_id
is the alternate, but the server side name it's associated with and
stored as is the alternate.

This was found while refactoring glance v2 for resource2.

Change-Id: Ib988da4826621982dc62c44ca905cd921afca7e8