Fix precedence for pass-in options
The passed-in argparse Namespace is flat and does not have an 'auth'
dict, all of the auth options are in the top level.  If the loaded
dict from clouds.yaml as an 'auth' dict, that used to totally override
any passed-in options.

This is backwards.

Make the passed-in auth options always win over clouds.yaml as this is
the only way to change/override those values from the command line.

Change-Id: Ic2752a396d45deeda19a16389437679829e0844d