Rebase orchestration to resource2/proxy2
This patch rebases the orchestration resources to resource2 and proxy2.
To make the orchestration service work, some methods from the base
resource2.Resource class are modified to allow passing of additional
parameters. These modifications are mainly due to the asymmetry of Heat
API design. We are trying to avoid overriding the whole method in
sub-classes in these changes.

Change-Id: I71372bb81351e0c83a1e013e50ba2c10cd984ed1