Allow creating a floating ip on an arbitrary port
There are use cases where creating a FIP on an unbound port is a
completely legitimate thing to do. Since our floating ip process already
starts with "hey man, find me a port please" - allowing a parameter
which short-circuits that and just attaches the FIP to the given port is
not too terrible, all things considered.

Change-Id: Iee2fc2ac7b26dfa2a48b0404a035e2820befa5ba