Add tests to show IP inference in missed conditions
There was no coverage concerning floating ip detection on clouds that
didn't have floating ips at all, or even the OS-EXT-IPS extension.
Although the problem was fixed in the previous commit, add a test that
shows it. Also, add a test for Rackspace specifically, since it's
network stack is different, to ensure that shade doesn't break dealling
with Infra's gerrit server.

Change-Id: I0bad8c69fe456a171d51f477dd6a5f8fb7065f60