Validate config vs reality better than length of list
Configuration indicating that a network is external only needs to
validate in one direction. Namely, if you indicate that network a is
external, and there exists a network be that can also be inferred to be
external, that's fine. But if you indicate network a is external and we
do not find network a, that's a problem.

The previous test using list length was woefully inadequate for this
purpose.

Change-Id: I3e7ece26de23007d1b97a7b8a25c3db31b124e5f