Allow image and flavor by name for create_server
We do a bunch of work everywhere else in shade to be friendly - but for
some reason we only accept object or id for image and flavor. Fix it.

In the unit tests, pass in a dict with an id value to avoid the need to
mock the glance client, since that's not really what we're testing in
any of those tests ... but add a test that does not do that to verify
that the glance and nova clients are used to look at image/flavor lists.

Change-Id: I1760a7464e43e19a475f6d277148a3c7e54ac468