Ensure per-resource caches work without global cache
With the move to dogpile based per-resource caches, we need to make sure
they still exist if they are configured, even if global cache isn't. To
do this, test to see if any are set, and if so, configure the global to
be FakeCache but the per-resource to be memory cache.

Change-Id: I8261b807352a4b54b19f6831b795eedc847e86ab