Rename _get_free_fixed_port to _nat_destination_port
The name _get_free_fixed_port is misleading. What the function actually
does is find a port attached to the server which is on a network which
has a subnet which can be a nat_destination. Such a network is referred
to in shade as a "nat_destination" network. So this then is a function
which returns a port on that network that is associated with this
server.

Change-Id: If69ad8f7c08a4df83bc00cb00801565a7473a9a5