Skip test creating provider network if one exists
The current default in devstack-gate for neutron is to create a public
provider network using the existing physical network. That's AWESOME and
we're highly in support of it - but it breaks our test of creating a
public provider network using the existing physical network.

Detect that this is the case and skip the test if it is.

Change-Id: Iff103154aace62f713b2f7b3303f19d31b20d0a7