Add ability to configure Session constructor
Sometimes it's useful to wrapp the keystoneauth.Session object. OSC has
a KeystoneSession object that injects timing data. shade is considering
one that wraps external calls in shade's TaskManager. Allow for passing
in a callable that will return a Session.

Almost no people will want to use this - it's a super advanced kind of
thing.

Change-Id: Ib64260916695e9fbea437862cd669a4fb85ec9e4