Use list_servers for polling rather than get_server_by_id
Using get_server_by_id bypasses the list_server poll/cache. There is one
legit place where we should do this, but in the other places it causes
us to poll with great frequency unnecessarily.

The TaskManager test changes are just to get test discover to stop
finding then skipping the fixtures in the file.

Change-Id: I7f1e360f53d7344dfc996eb4c4813184720f8da0