Generalize endpoint determination
Since moving away from the service catalog for per-service
endpoint determination, we've covered most of the cases we're familiar
with, but of course a cloud can configure endpoints in other ways,
including some we're not handling properly.

Rather than just use the root of the service catalog's entry, or
using the entire entry when it doesn't include a port, we need to
automate this further to cover a number of cases at once. We can't
make a smart determination on the root endpoint by just looking at the
URI anymore--we have to make some requests and see what happens.

This change now splits the URI and looks at the parts of the path and
reconstructs a number of sub-URIs. It first pops off any tenant/project
from the given URL, and then constructs successively longer paths
eminating from the root. We then make requests to each of them until we
find a URL which gives us the response we're aiming to parse.

Given a URL like http://mycloud.com/nova/v2/blahblahtenantid, we would
make the following requests while looking for a JSON body with a list of
versions:

1. http://mycloud.com/
2. http://mycloud.com/nova/ (this is the one that would work, and we
would exit)
3. http://mycloud.com/nova/v2

In some cases, services are configured either at a particular port or
under a subdomain. This was the original case we were solving for.
Given a URL like http://nova.mycloud.com/v2/blahblahtenantid, we would
make the following requests:

1. http://nova.mycloud.com/ (this is the one that would work, and we
would exit)
2. http://nova.mycloud.com/v2/

So, this does unfortunately add one or more extra GET requests to some
cases (one time on startup only), but there is no other automated way
to properly determine a root endpoint. The only other way we could
reliably know what root endpoint to go to would involve end-users
submitting some kind of parameter to a profile or other configuration in
order to specify what their environment looks like. They really should
not have to know that, and that complicates our user experience if we
need to enforce something like that.

Change-Id: I333638aee889202d8a17bf9e3df5c797c23a03e4