Provide better fallback when finding id values
At least in the case of object_store's Account resource, it
intentionally has no id or name value due to the way the REST API is
structured and how it maps to the concepts we're using. However, when
looking for the id value of an Account, we end up going through the
_alternate_id code and don't find one of those, and thusly end up trying
to look up an empty string as the alternate ID.

If we don't have an id or an alternate id, the id value should just be
None, so intercept any potential KeyError and just return None.

Change-Id: I60dde9f8f3c9a0eebaeadc8b2953136ad66c4ed4