Start using requests-mock for REST unit tests
Mocking out the whole client is the wrong level. Although we'd still
like to betamax the functional tests - as we put in REST calls, let's
make the mocking of them happen at the requests level so that we test as
much of shade as we can.

Nobody look at the v2 keystone catalog. The existing clouds.yaml entry
was for a v2 cloud and I wanted to keep the patch contained to the task
at hand. We should likely do some crazy testscenarios magic at some
point to make sure _all_ the things work with both v2 and v3.

Change-Id: Ife4d95df5417d329195c737814e89d1370a0597b