Add resource for DHCP Agent.
This patch introduces resource for DHCP agent in the OpenStack SDK. This
is used in implementing neutron DHCP agents commands in the OpenStack
client.

The equivalent neutron commands to be implemented in OSC:

- dhcp-agent-list-hosting-net
- dhcp-agent-network-add
- dhcp-agent-network-remove
- net-list-on-dhcp-agent

To implement the above commands in OSC, their respective resource in SDK
is needed. This patch introduces those.

Partially Implements: blueprint network-dhcp-adv-commands

Change-Id: Idebd3cd4b445a244962484923aa6cc47ebd8deed