Allow setting env variables for functional options
The shade functional tests are hardcoded to use the cloud configuration
that is typically provided by devstack. It turns out that shade
functional tests are a useful way to quickly test that a cloud is up and
configured.

Allow providing new values for flavor, image and cloud variables so that
the functional tests can be run on a non-devstack cloud.

Change-Id: I90ab125be77091eaf16ae16e47f4336d4ef49880