Logging: avoid string interpolation when not needed
Call to `format` or '%' (basically all string formating) can be
expensive and if someone is not interested in the logs produced
by shade (log_level set to error or higher) can be avoided.

Note: this change is a bit risky given the low unit test coverage.
But the longer we wait the harder this change will be.

Change-Id: I171f4ee08e5978c7b6e3853ed4ee533682c2ffe6