Update floating ip polling to account for DOWN status
The DT cloud transitions floating ips through the DOWN state as part of
normal operation. Don't ask me why. Put the check for port's not
matching after waiting for ACTIVE state to be hit, rather than before.
Also, don't throw an exception if the fip goes into DOWN while deleting.
Finally, skip adding floating ip info to a server from neutron if the
server is not in ACTIVE state, as a server is not expected to have valid
addresses when it's not in that state anyway.

Change-Id: I3020f31783f8d24cbd91942961a7c7868b51698d