Allow server-side filtering of Neutron floating IPs
Neutron API allows to filter floating IPs based on some fields. This
enables server-side (efficient) filtering that Shade should leverage
when it can.

Unfortunately Nova-net (which is deprecated) doesn't support server-side
filtering when it comes to FIP, so we have to distinguish the two at
several places.

This patch also updates the `search_floating_ips` method to use
server-side filtering when appropriate.

Change-Id: I28c960091c1679ca588c0cfb3bed1881cd6b03d2