Make search_projects a special case of list_projects
As the first step in aligning list and search, make list_projects have
all the things and search just be an alias. In this case, there were
some positional arguments already in place, so it can't be a 100%
straight alias - but it can at least only have one set of logic.

As part of this, added some infrastructure for dealing with filter logic
and pushdown conditions.

Change-Id: Iff349ff26e59c9ae8022e3ee8f0987197b0eccc4