Be specific about protected being bool
Protected needs to be bool - but other values need to get cast to
string. There's no good way other than enumerating them like we did for
the other values.

There are two ways to pass in user-defined settings - via the meta param
which does not get cast to anything else and assumes the values are in
the correct type, and **kwargs which does its best to cast.

We know about protected. We can add others if needed (although the
glance docs do not list any other boolean values)

Change-Id: Ib75246441e54d675987a18b890ce847043e7a80f