Remove test of having a thundering herd
The important test is the following one, which tests that we prevent
thundering herds. The test that we _have_ one is subject to a race
condition where the mocks execute so quickly that we don't herd even
without the protection. But that's ok - it's ok for a herd to
accidentally not happen. So remove the test which is just false-negative
noise.

Change-Id: I10c418a2c4ba2d8d1b0b3a22516bff2a688555a7