Add test to trap for missing services
Recently when there was an issue with the magnum devstack plugin causing
the shade gate to not have swift, we didn't notice except through the
ansible tests. That's because we have a bunch of has_service checks in
the tests themselves to deal with different configs. Unfortunately, that
makes it easy to fail open.

Put in a test, along with changes to devstack-gate jobs, to throw errors
if services do not show up that should.

Depends-On: I2433c7bced6c8ca785634056de45ddf624031509
Change-Id: I16f477c405583b315fff24929d6c7b2ca4f2eae3