Stop double-reporting extra_data in exceptions
We append extra_data to the args we send to Exception, but we also
append extra_data to the str output. Thing is - Exception already does
that with all the args - so we wind up with something like:

shade.exc.OpenStackCloudException: ("Image creation failed: <type
'exceptions.OSError'>", Munch({u'status': u'failure'...)) (Extra:
Munch({u'status': u'failure'...)

Also - the Munch output isn't terribly useful in the exception, so go
ahead and dict-ify it if we've passed a munch to extra_data.

Change-Id: Ie3271093fba0ac6f074e1cfd4e6a79455dc82be4