Clear the exception stack when we catch and continue
When we don't care about an exception in one of the poll loops, but then
there is a different unrelated error later on, the caught and handled
exception can linger. This is seen in output like this:

shade.exc.OpenStackCloudException: ("Image creation failed: <type
'exceptions.OSError'>", Munch({u'status': u'failure',...)) (Extra:
Munch({u'status': u'failure', ...)) (Inner Exception: (503) Server
Error: Service Unavailable for url: ...)

The 503 error is caught and skipped as a normal matter of course. Then
the task later fails and returns a status: failed, which we throw a new
exception for - but the earlier 503 error shows up in the exc_info call
so it gets reported. Since the 503 ISN'T related to the actual error
we're reporting, the output is erroenous and confusing.

Change-Id: I87b521df98435d541d62269c0987313add74d260