Switch swift calls to REST
Added support for calls that directly return lists to the adapter.

There are a few minor changes in the test suite. swiftclient appended
query strings differently than how requests is doing it. A comment was
adding clarifying why we're testing a 409. And the b'test body'
workaround for python3 issue we saw with request_mock and swiftclient
does not occur now, so we removed the b''.

Change-Id: I7e962f99cbab07ed1d4fea6e2c6df10776a4be80