Log request ids when debug logging is enabled
If debug logging is enabled, it's fairly likely that's because a user
wants to debug things. Often times having the request id in the log is a
useful part of that.

Also, while we're in there, don't try to log information about the
object returned when there is no object to query. And log a useful
message on lists, and when we get objects from nova.

Change-Id: I02579227f3475a952006689182f6ca112fa1f7ed