Add helper script to install branch tips
shade has a functional test that intends to test shade against the
current tip of client libs. Unfortunately, what it's really doing is
installing latest release of the library into the shade virtualenv
that's used for functional testing and using tip of library for the
OpenStack install.

This is not a combo we care about.

Instead, install the library tips into the virtualenv. To do this, make
a functional-tips venv for tox, and make the post_test_hook optionally
use it.

Change-Id: Ibac2982e88439362c7af933c3a07c4d581ee6f2b