Add raw client constructors for all the things
We want to start taking new functionality in with raw REST. However,
before we point people at doing that, we kind of need for there to be
clients people can use to make REST calls.

Go ahead and add everything os-client-config has default api versions
for.

Note - we could almost certainly get fancy with some sort of
metaprogramming to avoid the copy-pasta here - but I do not think the
complexity is worth it - it's not that much pasta.

Change-Id: I4691bc7c4943136b1e3fb9a9a8981e682b42ec8a