Copy in needed template processing utils from heatclient
There is a LOT of client-side file processing that goes on before
uploading heat templates to the cloud. Rather than reimplement it all
from scratch, copy in the relevant bits. With this done, removing
heatclient itself should be fairly easy. This will also allow us to fix
a long-standing bug, which is that heat event polling bypassed the
TaskManager since the utils make their own API calls.

Depends-On: Id608025d610de2099d7be37dcff35de33c10b9d5
Change-Id: I384f81b6198f874e78a434515123f955017e0172