Enforce inclusion of pulic proxy methods in docs
Per the approach on https://review.openstack.org/#/c/428276/, and in
general even outside of that change, we should have more strict
enforcement on the inclusion of our public proxy APIs in the
documentation. This tool runs as a part of our doc build, via `tox -e
docs`, and collects all of the public proxy methods and then compares
that list to the list of methods that were produced by Sphinx.

For right now this will only output warnings, but we should eventually
turn on enforcer_warnings_as_errors in doc/source/conf.py so that we can
truly enforce these things. If you don't document something, it'll kill
the doc build and then kill the docs gate job, so undocumented code
won't be accepted. We'll need to leave it off for the time
being as we transition into it.

Change-Id: I96743de7e0790da98d758415e084a26a92aa3c70