Pass task to post_task_run hook
It's hard for consumers to do anything useful in post_task_run without a
copy of the task. Also - not adding this as an optional becaues the
consumer this was written for, nodepool, has not started using it yet,
and I find it VERY hard to believe anyone else will have picked up on
this.

Change-Id: Ib48fe54feb8b6ed10da3e5fa9accf985e4b22623