Add support for indicating required floating IPs
Some clouds require that users add a floating IP to a server if the user
wants that server to be able to talk to things that are not on the
cloud. Some clouds do not require this and instead give servers a
directly attached IP. The only way a user can know is to boot a server,
then ask neutron for the port associated with that server, then find the
network the port came from and then try to infer whether or not that
network has the ability to route packets northbound. Of course, networks
don't actually communicate that quality directly, (router:external
doesn't mean a network routes externally, it means the network can have
a router attached to it to provide floating ips) so it's still hit and
miss.

Where we can, save the user the stress and strain of not knowing how
their cloud wants them to get an externally routable IP.

Change-Id: I1baf804ce28bc1997b2347c4648c5cc56c750ead