Rename demo_cloud to user_cloud
The operator cloud is "operator_cloud" which is clear. Even though the
devstack user is called "demo" - the point of the non-operator cloud is
that it's an end-user connection. Rename demo to user to remove any
confusion/ambiguity. (also, it's the same number of letters so pep8
doesn't get borked)

Change-Id: Icfd2dd1687099798ed2cdf9349cc5b7d27dade58