Wait for volumes to detach before deleting them
In our functional tests, we keep hitting errors with deleting volumes.
It would stand to reason that the problem is that we try to delete them
too quickly after deleting the server so cinder doens't yet know they're
detached. Put in a wait loop to try to get to a state where the volume
can be deleted.

If this works, it may be worth adding a method to shade itself to expose
this - or alternatively a wait_for_detach flag to delete server that
would not return until volumes are also detached. Or to add a
"delete_volumes" flag to delete server. However, deleting volumes
involves deleting data, so maybe it's fine.

Change-Id: Ic460b6bc7f65e8761b2d765b88502ef6b5a79973