Deprecate "wait_for" methods on ProxyBase
The wait_for_status and wait_for_delete methods on ProxyBase end up
attached to every service-specific proxy, though they're not generally
applicable. The proxy methods as they are just simply expose what's
available on the resource base class, so any direct users in a proxy
subclass should instead implement this on their own when it's needed.

Currently, only compute uses this kind of functionality, and it rightly
implements it itself by going to the resource2 module and calling the
module-level functions. Internally, we do have a few uses of the
wait_for methods attached to proxies, though they're all in functional
tests (for block_store, compute, and orchestration). We should consider
exposing those methods directly on those proxies rather than exposing
the methods through the base class to dozens of services that don't have
those needs.

These should be removed for 1.0. They're also currently generating
warnings on every service's documentation via the enforcer tool for not
being documented (which was how this was originally found).

Change-Id: I4733161959ab282915f1cc2956245a377417fced