Fix several concurrent shade gate issues
This is a big patch because there are more than one issue happening at
the same time and we have to fix all of them to fix any of them.

Force nova microversion to 2.0

The current use of novaclient is to get the latest microversion. So far
this has not been a problem, as shade deals with different payloads
across clouds all the time. However, the latest microversion to nova
broke shade's expectations about how usage reports work. Actual
microversion support is coming soon to shade, but is too much of a task
for a gate fix. In the meantime, pin to 2.0 which is available on all of
the clouds.

Produce some debug details about nova usage objects
Capture novaclient debug logging

In chasing down the usage issue, we were missing the REST interactions
we needed to be effective in chasing down the problem. novaclient passes
its own logger to keystoneauth Session, so we needed to include it in
the debug logging setup.

Also, add a helper function to make adding things like this easier.

Consume cirros qcow2 image if it's there

The move from ami to qcow2 for cirros broke shade's finding of it as a
candidate image. Move pick_image into the base class so that we can include
add_on_exception and error messages everywhere consistently.

Add image list to debug output on failure.
When we can't find a sensible image, add the list of images to the test
output so that we can examine them.

Change-Id: Ifae65e6cdf48921eaa379b803913277affbfe22a