Privatize session instance on Proxy subclasses
Each Proxy subclass has a session, but it's currently a public name, and
as such is accessible. Making it accessible is sort of a low risk thing,
but isn't really necessary—if someone needs a session instance they will
perhaps be creating their own which they pass into a Connection, or they
would be more likely to get the session off of a Connection instance
(which is fine being public there, I think).

This came up in the proxy doc work I'm doing, where every proxy has a
warning for not documenting the `session` name. At the proxy level it
really should just be a private name and a part of the implementation.

This change basically does a s/self.session/self._session/

Change-Id: Iccb423aad801115d593a6ad5ead408f48054c5d4