Fix the network service provider test
This test was pretty lame before since it just verified that
the result was a string.  Now it verifies that at least one
service provider exists and I think I picked one that should
be aroudn for a while.  The test failure message also prints
the list of providers now so it should be easier to debug.

Partial-bug: #1665495

Change-Id: Ief96558770d81dca81091e235b8d370883b14b94