Update devstack config to point to a valid image
The devstack config we link to in the docs has been out of date for
quite some time. I usually just comment out the IMAGE_URLS line so
it doesn't even bother, but we should actually update it. The change
falls in line with what the Heat docs recommend, and it's because of
Heat that we have this configuration line in the first place.

See
https://docs.openstack.org/developer/heat/getting_started/on_devstack.html
for their recommended config.

Change-Id: I3f55c576f0e8aefa776e40509f8b7b098665ff39