Don't glob match name_or_id
I was trying to bring up a host on vexxhost using the image name
"Ubuntu 16.04.1 LTS [2017-03-02]".  As described in the change,
because this gets run through fnmatch() directly it makes everything
barf as "[2017-03-02]" gets matched as an incorrect glob character
class.

This is a little tricky because all the external list_*, search_*,
get_* API functions basically pass name_or_id through directly.  None
of them have ever explicitly said that you could use globbing,
although it probably would have worked.  Any such use would have been
relying on implementation details of the _util functions, and I feel
like the fact it's _underscore_utils means that you probably had fair
warning not to do that.  And obviously the "filters" argument is meant
for this type of filtering.

Additionally, fnmatch() says that it normalises case depending on OS
filename case-sensitivity, which seems like incorrect behaviour for
these functions which are not actually related to filenames.

Thus I believe the matching should be just be a straight string
comparison.

I have added a test for this, but we also have to remove the tests for
globbing behaviour as they are incorrect now.

Change-Id: Id89bbf6306adebe96c1db417a4b8b5ae51023af0