Use data when the request has a non-json content type
Using the json= argument to the request call when the content/type is not
intended to be application/json is crossing wires, and in those cases we
should call data=json.dumps(). This actually breaks older versions of
keystoneauth, which are otherwise perfectly fine. For newer, it works
either way, but maybe let's just learn to send json payloads as data in
any case where the content-type isn't application/json.

Change-Id: Ie93840625e1418eaffa442fb0d93080040d2a06d